Component-Scoped State
This is the kind of difference between what React does vs what Solid can do. The previous isn't possible in React, and this commit is what React would do. You pass the state through the function's props instead, and that is what would dictate where the state changes were. Personally I think the previous version is what is easier to migrate to when updating an existing web app to a framework.